% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_course_porte.R
\name{create_gate_df}
\alias{create_gate_df}
\title{Create a DataFrame of gate results from a race}
\usage{
create_gate_df(db)
}
\arguments{
\item{db}{A data frame containing race data, including a column \code{Portes} with gate results as strings and a column \code{Embarcation}.}
}
\value{
A data frame with individual columns for each gate result, named "Porte_1", "Porte_2", etc., and an additional \code{Embarcation} column.
}
\description{
This function processes a column of gate results from the input data frame (\code{db}) and
creates a new data frame where each gate result is split into individual columns.
The columns are named "Porte_1", "Porte_2", etc., and contain numeric values.
The new data frame also includes the \code{Embarcation} column from the original data frame.
}
