% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_course_time.R
\name{plot_course_time}
\alias{plot_course_time}
\title{Plot the distribution of athletes based on race results}
\usage{
plot_course_time(db, athlete_name, group_all = FALSE)
}
\arguments{
\item{db}{A data frame containing race results, including columns like \code{Resultats}, \code{Nom}, \code{Embarcation}, \code{Temps_pur}, and \code{Penalites}.}

\item{athlete_name}{A vector of athlete names to highlight in the plot.}

\item{group_all}{A boolean indicating whether to group all athletes together or to facet by \code{Embarcation}.}
}
\value{
A ggplot object visualizing the distribution of athletes based on their race results.
}
\description{
This function visualizes the distribution of athletes based on their race results.
A histogram is generated to represent the number of athletes per second, with vertical lines
indicating the positions of the athletes specified in \code{athlete_name}. Athletes with results
exceeding a threshold are counted and optionally labeled on the plot.
}
